Move SoundManager to components.
